ia_vision/requirements.txt
tensorflow
torch
torchvision
numpy
pandas
matplotlib
opencv-python
scikit-learn
Pillow
jupyter
seaborn
